requirements.txt
google-generativeai>=0.3.0
pysrt>=1.1.2
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
requests>=2.31.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fastapi>=0.115.0
uvicorn>=0.30.0
datasets[audio]>=2.14.0
soundfile>=0.12.0
librosa>=0.10.0